Se crearon los propios archivos de los models para tener un proyecto más organizado
